+++ src/code42cli/sdk_client.py
@@ -2,7 +2,9 @@ import py42.sdk
 import py42.settings
 import py42.settings.debug as debug
 import requests
+from click import prompt
 from click import secho
+from py42.exceptions import Py42MFARequiredError
 from py42.exceptions import Py42UnauthorizedError
 from requests.exceptions import ConnectionError
 
@@ -15,7 +17,7 @@ py42.settings.items_per_page = 500
 logger = get_main_cli_logger()
 
 
-def create_sdk(profile, is_debug_mode):
+def create_sdk(profile, is_debug_mode, totp=None):
     if is_debug_mode:
         py42.settings.debug.level = debug.DEBUG
     if profile.ignore_ssl_errors == "True":
@@ -30,18 +32,24 @@ def create_sdk(profile, is_debug_mode):
         )
         py42.settings.verify_ssl_certs = False
     password = profile.get_password()
-    return validate_connection(profile.authority_url, profile.username, password)
+    return validate_connection(profile.authority_url, profile.username, password, totp)
 
 
-def validate_connection(authority_url, username, password):
+def validate_connection(authority_url, username, password, totp=None):
     try:
-        return py42.sdk.from_local_account(authority_url, username, password)
+        return py42.sdk.from_local_account(authority_url, username, password, totp=totp)
     except ConnectionError as err:
         logger.log_error(str(err))
-        raise LoggedCLIError("Problem connecting to {}".format(authority_url))
+        raise LoggedCLIError(f"Problem connecting to {authority_url}.")
+    except Py42MFARequiredError:
+        totp = prompt("Multi-factor authentication required. Enter TOTP", type=int)
+        return validate_connection(authority_url, username, password, totp)
     except Py42UnauthorizedError as err:
         logger.log_error(str(err))
-        raise Code42CLIError("Invalid credentials for user {}".format(username))
+        if "INVALID_TIME_BASED_ONE_TIME_PASSWORD" in err.response.text:
+            raise Code42CLIError(f"Invalid TOTP token for user {username}.")
+        else:
+            raise Code42CLIError(f"Invalid credentials for user {username}.")
     except Exception as err:
         logger.log_error(str(err))
         raise LoggedCLIError("Unknown problem validating connection.")